refactor(cli): move ServeArgs, bind gates, and run_serve out of main
The serve flag bag, loopback bind checks, and server boot path now live
in serve.rs. main.rs keeps the clap dispatcher and the other run_*
commands. Tests reach bind helpers via the serve module.

Constraint: no CLI or bind-policy behavior change
Rejected: also move build_limits / build_server_config | they are shared by other commands
Confidence: high
Scope-risk: moderate
